PR changes: function clarifications, env vars, gitignore, entrypoint
Moved all aggregator related functions into aggregator as static
functions.

Fixed env variables to match standard GOOGLE env variables

Editted .gitignore typo

Deleted softlink in Dockerfile and adjusted entrypoint call
