+++ .gitignore
@@ -3,7 +3,7 @@
 test.out
 .vscode/
 */*.swp
-coverage-*.htmli
+coverage-*.html
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
