+++ aggregator/sample.env
@@ -1,4 +1,4 @@
-GOOGLE_CREDENTIALS_FILE=path/to/gcloud/service/credentials.file
+GOOGLE_APPLICATION_CREDENTIALS=/path/to/gcloud/service/credentials.file
 AGGREGATION_CKPTS_FILELIST=sample-ckpts-filelist.txt
 AGGREGATION_TYPE=cma
 AGGREGATION_OUTPUT_PATH=gs://path/to/aggregation/output
